Change on the semantics of 'open M' and '(*@ open M *)'.
The 'open M' and '(*@ open M *)' will bring M into the current scope.
If 'M' is not in the current namespace, then we look for file named 'm.mli'.

If 'M.t' occurs in a type or 'M.x' occurs in an expression and 'M' is not
in the current namespace, we will try to load file 'm.mli' and open it in
the current context.