Consider comments as spaces while preprocessing
In the following specification:

```
(*@ type t *)
(*@ ephemeral *)
```

the preprocessor needs to recognize that `ephemeral` must be attached in
a nested attribute to the first attribute containing `type t`. It does
so by accepting `Space` between the two items, but not simple comments.

So this turn simple comments into another form of `Space` so that

```
(*@ type t *) (* a ghost type *)
(*@ ephemeral *)
```

is valid syntax.