Fix printing of mixfix symbols
This commit proposes support for printing partial and complete
application of mixfix symbols. Note that there is no support for mixfix
symbols with an arity greater than the number of underscore + 1 as there
are no such mixfix symbols defined in Gospel and there is no way at the
moment for the user to defined new mixfix symbols.